requirements.txt

py-solc==3.2.0
python-dotenv==0.12.0
web3==5.6.0

